refactor: extract SQLAlchemy models into models.py
Move ServiceEntry, Widget, WidgetValue, Group, and User out of
app.py and into a dedicated models.py. app.py now imports them
back. No schema or behavior change.

alembic/env.py now imports models instead of app to populate
db.metadata. The fallback note about the future move is removed
since this commit lands it.

Drops unused imports from app.py: UserMixin, generate_password_hash,
column_property, select, func.
